確認しました

+++ app/db/migrate/20210117114157_create_vtubers.rb
@@ -2,7 +2,7 @@ class CreateVtubers < ActiveRecord::Migration[5.2]
   def change
     create_table :vtubers do |t|
       t.string :name
-      t.integer :subscribers
+      t.float :subscribers
 
       t.timestamps
     end
